requirements/base.txt
channels-redis==4.0.0
daphne==4.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-oauth-toolkit==2.2.0
djangorestframework-simplejwt==5.2.2
inbox.py==0.0.6
loguru==0.6.0
pip-chill==1.0.1
psycopg2-binary==2.9.5
psycopg2-binary==2.9.5
pyopenssl==22.1.0
python-social-auth==0.3.6
sendgrid==6.9.7
service-identity==21.1.0
social-auth-app-django-mongoengine==1.0.0
pillow==9.4.0
django-webpush==0.3.4
